+++ CHANGELOG.md
@@ -11,3 +11,6 @@
   * Idea tags changed
   * Idea status changed (closed, opened)
 * Comments are now loaded from oldest
+* Board settings changed in admin panel now are updated properly for the whole page
+* Social link and webhooks creators won't break anymore when putting invalid url value
+* Replaced custom edited snarkdown library with marked for better markdown parsing and to fix broken parsing issues